hardware_v1: add initial prototype hardware
This is a pi HAT (not officially: no EEPROM, wrong size) that
has a DB25 on it and three 74AHCT541 8-channel buffers for
3V3/5V level conversion.  Pi GPIOs are mapped directly to
parallel port data, status, and control lines.  It should be
adequate for SPP mode.

It's layed out as a single sided board since that's easier for
me to prototype with for the first try.