dts: add device tree overlay
Add initial device tree overlay for parport-gpio module
to identify and provide initial config for GPIO pins assigned
to data, status, and control ports.

The "compatible" name links the driver and the dts file.

On the pi, after the compiled dtbo file is installed,
add the following to /boot/config.txt:

dtoverlay=parport-gpio