hardware_v1: status 10k pd -> 4.3k pu
Change the 10K pulldowns to 4.3K pullups.
4.7K is what would normally be used here but I had some
803 4.3K's on hand.

Place some vias for the pullups and also for the unrouted
CONTROL3 line.  Before we had a one sided board which is
a bit easier to fab with the CNC.  Place vias so it can
still be made one sided without requiring connectors to
have plated through holes.  A small number of jumper wires
will need to be soldered to vias on aone sided board.