driver: fix NULL oops
If CONFIG_PARPORT_NOT_PC is not defined and CONFIG_PARPORT_1284
is defined, internal 1284 "daisy" probing causes a NULL pointer
dereference (oops) at insmod of parport_gpio.ko.

This is likely due to a conditional in parport.h that, if NOT_PC
is undefined, hardwires operations to the PC versions, bypassing
the operations table.  I did not run down why the hardwired
symbols seem not to be an obstacle at parport.ko load time.

Fixes #1.