driver: delete port on module unload
Call parport_del_port() instead of parport_put_port()
when the module is unloaded, so that parport0 can be
reused if the module is re-loaded.  Otherwise we get
parport1, 2, and so on.

Also, change printk code so the port number is displayed
with the assigned GPIO pins e.g.

parport0: data on gpio pins [22,23,24,10,25,9,8,11]
parport0: status on gpio pins [18,17,4,3,2]
parport0: control on gpio pins [26,19,6,13]

Fixes #7