hardware_v3: rework v2 in KiCAD
Create a new hardware version with the following changes
1) Use KiCAD instead of Eagle so that other people can
more easily modify/remix.
2) Add HAT EEPROM to conform to Pi Foundation specs
3) OSHW logo on board

After I sent this to OSHPARK I realized I omitted the
bypass cap that v2 had on the buffer chip.  Board works,
so I'm assuming it's not a big deal.

Used the Pi HAT template included with KiCAD for the
board outline, Pi connector placement, and EEPROM.