travis-ci: build kernel module in various environments
Problem: travis-ci is not building the kernel module.
It's only running checkpatch.pl on it.

Add a build matrix to cover linux-4.4 and linux-4.14
on amd64 and arm64.

Don't build the parport subdirectory as it is already
included in Ubuntu kernel builds, and is not very
portable across kernel versions.