v5: import/reimplement parport design on template
Problem: template board is still blank.

Adjust board edge to handle SMT header and add cutout for pi5 camera/
display FFC connectors.  For GPIO connector, use existing kicad 8 Harwin
M20 series footprint but adjust the pinout for top of board pi GPIO,
and adjust position since centroid moved from THT pin 1 to SMT centroid.
Add keepout for pi5 PoE like the existing one for pi4.

Import the SN74LVC16128 symbol from the v4 board.  Orient the chip
as before for best routing but aim for tighter traces this time.
Add bypass caps to all power inputs for good measure.  Chip capacitors
and resistors are shrunk from 0805 to 0603 in this revision.  Allow the
cable VCC to be jumpered to either 5V or 3V3 per the spec sheet.

Move the EEPROM to the east end of the board for more direct routing.
Add a strong 1K pull-up to the template's WP jumper per the HAT+ spec.
Orient the chip the same direction as the SN74LVC16128 to avoid confusion.

Add the DB-25 right-angle 2x13 header on the west end as before.
This time, use a custom header footprint so the DB-25 pin numbers can
be mirrored on the header.  Also, add a board cutout to improve insertion
ergonmics if the right angle header is used.

Implement the GPIO reassignments proposed in #67.  This should allow
I2C, UART0, and PWM0 to be used in stacked HATs.

Add a Sparkfun Qwiic connector to allow I2C expansion that way as well.
See: https://www.sparkfun.com/qwiic.

Put the official HAT+ logo on the board since this board complies with
https://datasheets.raspberrypi.com/hat/hat-plus-specification.pdf.
The logo image was provided via email along with marking guidelines.
Use the minimum 2.25mm height specified in the guidelines.

Add some helpful notes and the DB-25 pinout to the board silkscreen.